+++ src/format/w3c_vc/store.rs
@@ -10,7 +10,7 @@ use serde_json::Value;
 use crate::core::Kind;
 use crate::format::FormatProfile;
 use crate::format::w3c_vc::{VerifiableCredential, W3cVcClaims};
-use crate::oid4vci::types::CredentialDefinition;
+use crate::oid4vci::issuer::CredentialDefinition;
 use crate::oid4vp::verifier::{Claim, Queryable};
 
 /// Convert a `w3c` credential to a `Queryable` object.
