+++ src/oid4vci/handlers/authorize.rs
@@ -76,7 +76,7 @@ use crate::oid4vci::error::{invalid, server};
 use crate::oid4vci::handlers::{Body, Error, Handler, Request, Response, Result};
 use crate::oid4vci::provider::{Metadata, Provider, StateStore, Subject};
 use crate::oid4vci::state::{Authorized, Expire, Offered};
-use crate::oid4vci::types::{
+use crate::oid4vci::issuer::{
     AuthorizationCredential, AuthorizationDetail, AuthorizationDetailType, AuthorizationRequest,
     AuthorizationResponse, AuthorizedDetail, Issuer, RequestObject,
 };
