+++ src/oid4vci/handlers/create_offer.rs
@@ -17,7 +17,7 @@ use crate::oid4vci::error::{invalid, server};
 use crate::oid4vci::handlers::{Body, Error, Handler, Request, Response, Result};
 use crate::oid4vci::provider::{Metadata, Provider, StateStore, Subject};
 use crate::oid4vci::state::{Expire, Offered};
-use crate::oid4vci::types::{
+use crate::oid4vci::issuer::{
     AuthorizationCodeGrant, AuthorizationCredential, AuthorizationDetail, AuthorizationDetailType,
     AuthorizedDetail, CreateOfferRequest, CreateOfferResponse, CredentialOffer, Grants, Issuer,
     OfferType, PreAuthorizedCodeGrant, SendType, Server, TxCode,
