+++ src/oid4vci/handlers/deferred.rs
@@ -11,11 +11,11 @@
 use anyhow::Context as _;
 
 use crate::oid4vci::error::invalid;
+use crate::oid4vci::handlers::credential::credential;
 use crate::oid4vci::handlers::{Body, Error, Handler, Request, Response, Result};
-use crate::oid4vci::issuer::credential::credential;
 use crate::oid4vci::provider::{Provider, StateStore};
 use crate::oid4vci::state::Deferred;
-use crate::oid4vci::types::{
+use crate::oid4vci::issuer::{
     CredentialHeaders, CredentialResponse, DeferredCredentialRequest, DeferredCredentialResponse,
     DeferredHeaders,
 };
