+++ src/oid4vci/issuer/offer.rs
@@ -8,7 +8,7 @@ use serde::{Deserialize, Serialize};
 
 use crate::core::urlencode;
 use crate::oauth::GrantType;
-use crate::oid4vci::types::{AuthorizationCodeGrant, Grants, PreAuthorizedCodeGrant};
+use crate::oid4vci::issuer::{AuthorizationCodeGrant, Grants, PreAuthorizedCodeGrant};
 
 /// Request a Credential Offer for a Credential Issuer.
 #[derive(Clone, Default, Debug, Deserialize, Serialize)]
