+++ src/oid4vci/provider.rs
@@ -15,7 +15,7 @@ use anyhow::{Result, anyhow};
 use credibil_identity::{IdentityResolver, SignerExt};
 
 use crate::blockstore::BlockStore;
-use crate::oid4vci::types::{Client, Dataset, Issuer, Server};
+use crate::oid4vci::issuer::{Client, Dataset, Issuer, Server};
 pub use crate::state::StateStore;
 use crate::token_status::StatusStore;
