--- src/oid4vci/server.rs
@@ -1,9 +1,0 @@
-//! # Authorization Server
-//!
-//! The `OAuth2` API protecting the credential issuer.
-
-pub mod authorize;
-pub mod metadata;
-pub mod par;
-pub mod register;
-pub mod token;
